Fix color_code type mismatch and unchecked DMR dynamic_cast
- Unify Call_Data_t::color_code from long to unsigned int to match
  Transmission::color_code (color codes are 0-15, never negative)
- Add null check on dynamic_pointer_cast<dmr_recorder> in
  conclude_call() to prevent nullptr dereference

+++ trunk-recorder/global_structs.h
@@ -92,7 +92,7 @@ enum Recorder_Type { DEBUG,
 
 struct Call_Data_t {
   long talkgroup;
-  long color_code;
+  unsigned int color_code;
   std::vector<unsigned long> patched_talkgroups;
   std::string talkgroup_tag;
   std::string talkgroup_alpha_tag;
